Smooth colored lightning. (Lantern glows yellow, player glows red.. there are a couple of bugs when 2 colors/intensity collide but otherwise it's ok) Btw 3x3 tiles around the player get deleted (to test vertexarray regeneration)

+++ src/app/states/StateMenu.cpp
@@ -19,7 +19,7 @@ namespace State
 	{
 		m_backGround.setTexture(&app->getResources().textures.get("gui/menu/dev/background"));
 
-		m_devBuildLabel.setPosition(Vec2(18, app->getWindow().getSize().y - 40));
+		m_devBuildLabel.setPosition(Vec2(18.f, static_cast<float>(app->getWindow().getSize().y) - 40.f));
 		m_devBuildLabel.getText().setFillColor(sf::Color(255, 235, 0, 255));
 
 		m_UI.addComponent(m_singleplayer);
